Updates, adds .gzip reading
