+++ cmd/root.go
@@ -3,12 +3,12 @@ package cmd
 import (
 	"os"
 
-	"github.com/bmeg/jsonschemagraph/cmd/gengraph"
+	"github.com/bmeg/jsonschemagraph/cmd/data_validate"
+	"github.com/bmeg/jsonschemagraph/cmd/gen_dir"
+	"github.com/bmeg/jsonschemagraph/cmd/gen_graph"
 	"github.com/bmeg/jsonschemagraph/cmd/schema_graph"
 	"github.com/bmeg/jsonschemagraph/cmd/schema_lint"
 	"github.com/spf13/cobra"
-	"github.com/bmeg/jsonschemagraph/cmd/data_validate"
-
 )
 
 // RootCmd represents the root command
@@ -19,7 +19,8 @@ var RootCmd = &cobra.Command{
 }
 
 func init() {
-	RootCmd.AddCommand(gengraph.Cmd)
+	RootCmd.AddCommand(gen_dir.Cmd)
+	RootCmd.AddCommand(gen_graph.Cmd)
 	RootCmd.AddCommand(schema_lint.Cmd)
 	RootCmd.AddCommand(schema_graph.Cmd)
 	RootCmd.AddCommand(data_validate.Cmd)
