+++ cmd/schema_lint/main.go
@@ -4,7 +4,7 @@ import (
 	"fmt"
 
 	"github.com/bmeg/grip/log"
-	jsgraph "github.com/bmeg/jsonschemagraph/util"
+	"github.com/bmeg/jsonschemagraph/util"
 	"github.com/spf13/cobra"
 )
 
@@ -14,7 +14,7 @@ var Cmd = &cobra.Command{
 	Args:  cobra.MinimumNArgs(1),
 	RunE: func(cmd *cobra.Command, args []string) error {
 
-		sch, err := jsgraph.Load(args[0])
+		sch, err := util.Load(args[0])
 		if err == nil {
 			for _, cls := range sch.Classes {
 				fmt.Printf("OK: %s (%s)\n", cls.Title, cls.Location)
